+++ src/api-parser-v2.ts
@@ -8,9 +8,9 @@ import { dirname, join } from 'pathe';
 import { compile } from 'stylis';
 
 import { apiv2 as userAgents } from '../data/user-agents.json';
-import type { APIResponse } from './index';
-import { APIDirect, APIv2 } from './index';
-import type { FontObjectV2 } from './types';
+import type { } from './api-gen';
+import { APIDirect, APIv2 } from './data';
+import type { APIResponse, FontObjectV2 } from './types';
 import { orderObject, weightListGen } from './utils';
 import { validate } from './validate';
