+++ src/validate.ts
@@ -1,10 +1,9 @@
 import consola from 'consola';
 import colors from 'picocolors';
 
-import type { FontObject } from './index';
-// eslint-disable-next-line import/no-cycle
-import { APIv1, APIv2, APIVariable } from './index';
+import { APIv1, APIv2, APIVariable } from './data';
 import { fontObjectValidate, fontObjectVariableValidate } from './schema';
+import type { FontObject } from './types';
 
 type Version = 'v1' | 'v2' | 'variable';
